refactor: move pure methods to utils
